Use the same i18n config for the three apps
