+++ src/feedback/translations/i18n.js
@@ -1,34 +1,14 @@
-import i18next from 'i18next'
-import LanguageDetector from 'i18next-browser-languagedetector'
 import enTranslation from './languages/en.feedback'
 import frTranslation from './languages/fr.feedback'
+import { newI18nInstance } from '../../utils/newI18nInstance'
 
-const i18n = i18next.createInstance()
-i18n.use(LanguageDetector).init({
-    // we init with resources
-    resources: {
-        en: {
-            translations: enTranslation,
-        },
-        fr: {
-            translations: frTranslation,
-        },
+const i18n = newI18nInstance({
+    en: {
+        translations: enTranslation,
     },
-    fallbackLng: 'en',
-    debug: process.env.NODE_ENV !== 'production',
-
-    // have a common namespace used around the full app
-    ns: ['translations'],
-    defaultNS: 'translations',
-
-    interpolation: {
-        escapeValue: false, // not needed for react!!
-        formatSeparator: ',',
-    },
-    react: {
-        useSuspense: false,
+    fr: {
+        translations: frTranslation,
     },
 })
 
-document.documentElement.lang = i18n.language
 export default i18n
